Move folder specific items out of menu component
