+++ frontend/components/FolderMenu/index.ts
@@ -1,5 +1,6 @@
 import FolderMenu from "./FolderMenu";
 import FolderMenuButton from "./FolderMenuButton";
-import FolderMenuListItem from "./FolderMenuListItem";
+import FolderMenuListItem, { MenuItem } from "./FolderMenuListItem";
 
 export { FolderMenu, FolderMenuButton, FolderMenuListItem };
+export type { MenuItem };
