+++ frontend/pages/workspaces/[workspaceId]/folders/[folderId]/index.tsx
@@ -5,8 +5,15 @@ import { useRouter } from "next/router";
 import styled from "styled-components";
 
 import { MobileFileList, FileTable } from "../../../../../components/FileTable";
-import { FolderMenu } from "../../../../../components/FolderMenu";
+import { FolderMenu, MenuItem } from "../../../../../components/FolderMenu";
 import { Head } from "../../../../../components/Head";
+import {
+  DeleteIcon,
+  EditIcon,
+  LockIcon,
+  MoveIcon,
+  UploadIcon,
+} from "../../../../../components/Icon";
 import { MainHeading } from "../../../../../components/MainHeading";
 import { NavHeader } from "../../../../../components/NavHeader";
 import { Navigation } from "../../../../../components/Navigation";
@@ -58,6 +65,34 @@ const FolderHomepage: NextPage = () => {
     variables: { folder: folderId },
   });
 
+  const items: MenuItem[] = [
+    {
+      title: "Upload file to this folder",
+      icon: <UploadIcon />,
+      href: `/workspaces/${workspaceId}/folders/${folderId}/upload-file`,
+    },
+    {
+      title: "Edit folder details",
+      icon: <EditIcon />,
+      href: "#",
+    },
+    {
+      title: "Move folder",
+      icon: <MoveIcon />,
+      href: "#",
+    },
+    {
+      title: "View folder permissions",
+      icon: <LockIcon />,
+      href: "#",
+    },
+    {
+      title: "Delete folder",
+      icon: <DeleteIcon />,
+      href: "#",
+    },
+  ];
+
   return (
     <>
       <Head
@@ -78,9 +113,7 @@ const FolderHomepage: NextPage = () => {
           <PageContent>
             <MainHeading
               withBorder
-              menu={
-                <FolderMenu workspaceId={workspaceId} folderId={folderId} />
-              }
+              menu={<FolderMenu folderId={folderId} items={items} />}
             >
               {folder.data?.folder.title || ""}
             </MainHeading>
